requirements.txt
flask
pytest
coverage
waitress
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
# opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
# opentelemetry-exporter-jaeger
requests
